Follow Java Naming Conventions
And removed unused imports and variables
